Update: Improve & Fix existing features
Update: The video background in the TWAOS page is now a continous randomized sequence with support for custom rules which can be implemented in the future.
Update: Move hover functionality for the Sip character on the TWAOS page in it's own function in the page's lib instead of the global one.